Move Actix Web 1 errors under actix_web_1 module
RestApiServerError and RequestError are both related to the Actix Web
1 implementation of the REST API, so move them accordingly.

This is intended to make it easier to move the REST API into another
crate in the future.
